requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2
joblib>=1.0.1
sqlalchemy>=1.4.23
openpyxl>=3.0.7
fastapi>=0.68.0
uvicorn>=0.15.0
plotly>=5.0.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
pydantic>=1.8.0
